Added Static GZIPping of (Textual) Resources Support
We now try to create static `gzip`s of all kinds of resources.
This is mainly interesting for textual resources (html, css, js), but
may potentially also work for pdfs and even images. If we gzip, e.g., an
image and the result should be smaller than the original, we keep the
compressed resource, otherwise we remove it.

With the right setup in your .htaccess file, you can then serve the
pre-compressed files, which will reduce the workload of your server
significantly.