Now using Ultra GZIP as Compressor
I have "outsourced" the compression of static resources with gzip to my
new "Ultra GZIP" project. This project has more functionality than the
old way to compress data. The compression is slower (I need to figure
out how much and whether it is still acceptable) but the results are
better too.